Objeto MySQLConnector esta implementado y funcionando, se ha añadido el objeto LectorDeArchivosCSV y efectivamente se puede leer un archivo de configuracion.

+++ src/main/java/edu/utn/ar/utils/Participante.java
@@ -1,22 +1,21 @@
 package edu.utn.ar.utils;
 
 public class Participante {
-
     private final Integer identificacionUnivoca;
     private final String nombre;
-    private Float puntosAcumulados;
+    private int puntosAcumulados;
     public Participante(Integer id, String nombre){
         this.identificacionUnivoca = id;
         this.nombre = nombre;
-        this.puntosAcumulados = 0.0f;
+        this.puntosAcumulados = 0;
     }
     public String getNombre() {
         return this.nombre;
     }
-    public Float getPuntosAcumulados(){
+    public int getPuntosAcumulados(){
         return this.puntosAcumulados;
     }
-    public void adicionarPuntos(Float p){
+    public void adicionarPuntos(int p){
         this.puntosAcumulados = this.getPuntosAcumulados() + p;
     }
     public Integer getIdentificacionUnivoca() {
